+++ protocols/identify/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.44.1 - unreleased
 
+- Support (enabled by default) pushing changes of external addresses.
+  See [PR 4885](https://github.com/libp2p/rust-libp2p/pull/4885).
 - Ensure `Multiaddr` handled and returned by `Behaviour` are `/p2p` terminated.
   See [PR 4596](https://github.com/libp2p/rust-libp2p/pull/4596).